Create Buy The Dip 1.0
https://usethinkscript.com/threads/buy-the-dip-v1-0-released-standard-version.3545/
https://youtu.be/kCKYb8z6hEM

Who is this indicator for?
Buy the Dip was designed specifically for swing trading. And as the name suggested, we will only be buying and holding stocks or options contracts, whichever you prefer.

To all the swing traders, this is going to be useful for you.

The standard version of this indicator comes in the form of an oscillator. We will be using it as a lower study. I will also post another version of the indicator where you can use it on the upper chart. That version will be released shortly after.

How do I use this indicator?
Ready? It's quite simple. When the green line reaches 0 or is below 0, we can expect a bottom. In other words, you can anticipate the pullback to be over, and the stock is ready to reverse.