Create Buy the Dip: Conservative Model
https://usethinkscript.com/threads/buy-the-dip-conservative-model.3762/
----------------------------------------------------------------------------------------------------
Here is the conservative model of our Buy the Dip indicator.

Why "conservative"?
The original BTD indicator uses the length value of 22. We strictly follow that length because it was provided by Larry Williams in his Synthetic VIX formula. And because there are about 22 trading days in a month.

While this does a fantastic job at identifying market bottoms, I think 22 as a length should be used for short term outlook. I created three different versions of our BTD indicator. Each with different length values. In this experiment, I replaced 22 with 50, 100, and 200. Why these three values? Let's just say the popular daily simple moving averages heavily inspired me.

Because 200 is such an extended period, at least for this indicator, I decided that we would not be using it. Therefore, 50 and 100 came out on top.

Takeaways
While I was only able to play around with this for a short period, I've learned that allowing all three different lengths to come together and meet at the 0 line increases the indicator's success rate. If you're more proactive, you may choose to wait for only 2 (the short and medium lengths).

If you decide to use the conservative model, do keep in mind that you will miss out on some great opportunities given by the Aggressive Model.

Conclusion
In conclusion, I want to point out that this is just an experiment. I wanted to share it here for you guys to play around with. Hopefully, learn something from using different length values to increase the probability.

By default, you are going to see two different lines.
Orange line = Medium term length = 50
Red line = Long term length = 100
If you go into the indicator's settings and enable the option called "Aggressive Model," you will see the third line, which is the default BTD line. I included the default length so that you don't have to display multiple versions of the indicator on your chart.

How to get the arrows to show up?

Go to your indicator's settings and select Yes for Arrows.

Why do I see down arrows?

Since the arrow may overlap when the medium and long term lengths cross below the 0 line simultaneously, I decided to use the up arrow for the medium line and down arrow for the long term line. I hope that makes sense.