Create Buy the Dip MTF (Multi Timeframe) Labels
https://usethinkscript.com/threads/buy-the-dip-mtf-multi-timeframe-labels.3600/

Labels that display Buy the Dip values from higher timeframes onto a lower timeframe chart.

Limitation: 1HR is the lowest timeframe from this script. So if you go to any timeframe higher than the 1HR, all the labels will disappear. I believe there is a solution to this. But I haven't been able to figure it out yet.

Alternative to MTF labels: One alternative to this MTF label script is to use the watchlist column.