Create Buy the Dip Indicator with Filtered Signals [Advanced Version]
https://usethinkscript.com/threads/buy-the-dip-indicator-with-filtered-signals-advanced-version.3549/

Here is the advanced version of the standard Buy the Dip indicator. It's no longer just an oscillator.

Main Features:
Solely act as an upper study
Filtered signals
Moving averages from the daily chart

How is this different from the standard version?
The lower study is no longer being utilized or displayed. With this version, we added the 50 and 200 DMA (simple moving average from the Daily chart) to filter out false signals.

In order to generate a signal, the indicator must satisfy the following condition:
Buy the Dip (green line) crosses below 0
The stock must be trading above the 50 or 200 Daily SMA (simple moving average)

⚠ Why the 50 and 200 SMA? If you know anything about long term trading, the 50 and 200 SMA can play a crucial role in identifying the trend of a stock. The stock is considered to be in an uptrend if it's trading above its 50 and/or 200 SMA. As a result, we should only be buying the dips if the stock is in an uptrend. Of course, this is NOT the golden rule. Some stocks can still offer great buying opportunities even when they're below their 50/200 SMA.

Minor Features
Ability to turn off the labels (Meter and MAs)
Ability to turn off the "Buy the Dip" bubbles
Ability to hide the moving average lines
Ability to adjust the timeframe of the moving average lines
Ability to switch between different type of moving averages (ex: Simple, Exponential, Wilders, and Hull)
Ability to adjust the length of the multiple averages (default: 50 and 200 SMA)